refactor : JPA 페이징 기능 적용을 위한 코드 수정
Related:#133
