+++ application/src/main/java/core/application/movies/repositories/mapper/CachedMovieMapper.java
@@ -27,6 +27,8 @@ public interface CachedMovieMapper {
 
 	List<CachedMovieEntity> findMoviesOnRatingDescendWithGenre(int offset, String genre);
 
+	int selectGenreMovieCount(String genre);
+
 	void update(@Param("movieId") String movieId, @Param("replacement") CachedMovieEntity replacement);
 
 	void delete(String movieId);
