+++ application/src/main/resources/mappers/movies/CommentMapper.xml
@@ -124,6 +124,11 @@
         from comment_table
     </select>
 
+    <select id="countByMovieId">
+        count (*) from comment_table
+        where movie_id = #{movieId}
+    </select>
+
     <update id="update">
         update comment_table
         set content=#{content},
